requirements.txt
apt-get install -y tesseract-ocr
streamlit
pandas
plotly
pillow
pytesseract
